+++ docs/source/api/fisher.rst
@@ -1,0 +1,9 @@
+Fisher Information Matrix (FIM) / IFVP
+=======
+.. autofunction:: dattri.func.fisher.ifvp_explicit
+.. autofunction:: dattri.func.fisher.ifvp_at_x_explicit
+
+.. autofunction:: dattri.func.fisher.ifvp_datainf
+.. autofunction:: dattri.func.fisher.ifvp_at_x_datainf
+
+.. autofunction:: dattri.func.fisher.ifvp_at_x_ekfac
