+++ docs/source/api/hessian.rst
@@ -1,0 +1,16 @@
+Hessian, HVP, and IHVP
+=======
+.. autofunction:: dattri.func.hessian.hvp
+.. autofunction:: dattri.func.hessian.hvp_at_x
+
+.. autofunction:: dattri.func.hessian.ihvp_explicit
+.. autofunction:: dattri.func.hessian.ihvp_at_x_explicit
+
+.. autofunction:: dattri.func.hessian.ihvp_cg
+.. autofunction:: dattri.func.hessian.ihvp_at_x_cg
+
+.. autofunction:: dattri.func.hessian.ihvp_arnoldi
+.. autofunction:: dattri.func.hessian.ihvp_at_x_arnoldi
+
+.. autofunction:: dattri.func.hessian.ihvp_lissa
+.. autofunction:: dattri.func.hessian.ihvp_at_x_lissa
