--- docs/source/api/hvp_ihvp.rst
@@ -1,21 +1,0 @@
-HVP/IHVP
-=======
-.. autofunction:: dattri.func.ihvp.hvp
-.. autofunction:: dattri.func.ihvp.hvp_at_x
-
-.. autofunction:: dattri.func.ihvp.ihvp_explicit
-.. autofunction:: dattri.func.ihvp.ihvp_at_x_explicit
-
-.. autofunction:: dattri.func.ihvp.ihvp_cg
-.. autofunction:: dattri.func.ihvp.ihvp_at_x_cg
-
-.. autofunction:: dattri.func.ihvp.ihvp_arnoldi
-.. autofunction:: dattri.func.ihvp.ihvp_at_x_arnoldi
-
-.. autofunction:: dattri.func.ihvp.ihvp_lissa
-.. autofunction:: dattri.func.ihvp.ihvp_at_x_lissa
-
-.. autofunction:: dattri.func.ihvp.ihvp_datainf
-.. autofunction:: dattri.func.ihvp.ihvp_at_x_datainf
-
-.. autofunction:: dattri.func.ihvp.ihvp_at_x_ekfac
